Move uptimerobot coordinator to its own file
